% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Meta-analysis workhorse function}
\usage{
meta_analysis(beta, se, hk = FALSE)
}
\arguments{
\item{beta}{Vector of effect sizes}

\item{se}{Vector of standard errors}

\item{hk}{Logical, whether Hartung-Knapp SEs should be calculated}
}
\value{

}
\description{
Meta-analysis workhorse function
}
